feat: remove has_permission chore: refactor utils ui: minor fixes
