+++ .gitignore
@@ -12,5 +12,5 @@ drive/www/drive.html
 .yarn-integrity
 **/__pycache__
 node_modules/
-experiments
+experiments/*
 *.sql
