+++ drive/drive/doctype/drive_file/drive_file.py
@@ -2,16 +2,17 @@ import frappe
 from frappe.model.document import Document
 from pathlib import Path
 import shutil
-from drive.utils.files import (
+
+from drive.utils import (
+    get_ancestors_of,
+    generate_upward_path,
     get_home_folder,
-    get_new_title,
-    get_team_thumbnails_directory,
     update_file_size,
-    FileManager,
 )
-from drive.api.files import get_ancestors_of
-from drive.utils.files import generate_upward_path
+from drive.utils.files import FileManager
 from drive.api.activity import create_new_activity_log
+from drive.api.files import get_new_title
+from drive.api.permissions import user_has_permission
 
 
 class DriveFile(Document):
@@ -35,12 +36,7 @@ class DriveFile(Document):
 
         if self.is_group or self.document:
             for child in self.get_children():
-                has_write_access = frappe.has_permission(
-                    doctype="Drive File",
-                    doc=self,
-                    ptype="write",
-                    user=frappe.session.user,
-                )
+                has_write_access = user_has_permission(self, "write")
                 child.delete(ignore_permissions=has_write_access)
 
     def after_delete(self):
@@ -146,12 +142,7 @@ class DriveFile(Document):
             parent_is_group = frappe.db.get_value("Drive File", new_parent, "is_group")
             if not parent_is_group:
                 raise NotADirectoryError()
-            if not frappe.has_permission(
-                doctype="Drive File",
-                doc=new_parent,
-                ptype="write",
-                user=frappe.session.user,
-            ):
+            if not user_has_permission(new_parent, "upload"):
                 frappe.throw(
                     "Cannot paste to this folder due to insufficient permissions",
                     frappe.PermissionError,
@@ -312,12 +303,8 @@ class DriveFile(Document):
         return self.name
 
     def permanent_delete(self):
-        write_access = frappe.has_permission(doctype="Drive File", doc=self, ptype="write")
-        parent_write_access = frappe.has_permission(
-            doctype="Drive File",
-            doc=self.parent_entity,
-            ptype="write",
-        )
+        write_access = user_has_permission(self, "write")
+        parent_write_access = user_has_permission(self.parent_entity, "write")
 
         if not (write_access or parent_write_access):
             frappe.throw("Not permitted", frappe.PermissionError)
@@ -349,12 +336,7 @@ class DriveFile(Document):
         :param share: 1 if share permission is to be granted. Defaults to 0
         """
         if frappe.session.user != self.owner:
-            if not frappe.has_permission(
-                doctype="Drive File",
-                doc=self,
-                ptype="share",
-                user=frappe.session.user,
-            ):
+            if not user_has_permission(self, "share"):
                 for owner in get_ancestors_of(self.name):
                     if frappe.session.user == frappe.get_value(
                         "Drive File", {"name": owner}, ["owner"]
