+++ drive/utils/users.py
@@ -6,17 +6,17 @@ import os
 
 
 def mark_as_viewed(entity):
-    if frappe.session.user == "Guest":
-        return
-    if not frappe.has_permission(
-        doctype="Drive Entity Log", ptype="write", user=frappe.session.user
+    if (
+        frappe.session.user == "Guest"
+        or not frappe.has_permission(
+            doctype="Drive Entity Log", ptype="write", user=frappe.session.user
+        )
+        or entity.is_group
     ):
         return
-    if entity.is_group:
-        return
-    entity_name = entity.name
+
     entity_log = frappe.db.get_value(
-        "Drive Entity Log", {"entity_name": entity_name, "user": frappe.session.user}
+        "Drive Entity Log", {"entity_name": entity.name, "user": frappe.session.user}
     )
     if entity_log:
         frappe.db.set_value(
@@ -24,7 +24,7 @@ def mark_as_viewed(entity):
         )
         return
     doc = frappe.new_doc("Drive Entity Log")
-    doc.entity_name = entity_name
+    doc.entity_name = entity.name
     doc.user = frappe.session.user
     doc.last_interaction = now()
     doc.insert()
